Object-storage-mock: Fix off-by-1 for HTTP range requests
The start and end byte locations of HTTP range request headers are _inclusive_.

If a HTTP Range header is present and specifies 0/9223372036854775807 for start/end, the calculation for the request-length in HeapStorageBucket's object range-request handling was wrong (off by one). As `Long.MAX_VALUE/9223372036854775807 - 0L + 1L` becomes `-1L`, a new function `Range.everything()` was introduced to prevent this overflow.