Quarkus: Tackle some more deprecated/changed configs
Tackles
```
Unrecognized configuration key "quarkus.log.sentry" was provided; it will be ignored; verify that the dependency extension for this configuration is set or that you did not make a typo
The "quarkus.swagger-ui.enable" config property is deprecated and should not be used anymore.
The "quarkus.smallrye-openapi.enable" config property is deprecated and should not be used anymore.
The "quarkus.log.file.json" config property is deprecated and should not be used anymore.
The "quarkus.log.console.json" config property is deprecated and should not be used anymore.
```