Bugfix: S3 remote signing for batch deletions does not work
The HTTP uri in the S3 sign request contains URIs with a path that only contains the bucket, but no path (like `http://localhost:12345/bucket?delete`). This fails the path-like check in `org.projectnessie.catalog.files.s3.S3Utils#asS3Location`.

This change updates the regex to fix the issue.