Cache: rework overshoot handling since Caffeine 3.2.3
There is sadly no fix possible with Caffeine since 3.2.3. See [this reply](https://github.com/ben-manes/caffeine/issues/1897#issuecomment-3481719258).

The workaround I came up with is to change the logic a bit:

* If the cache-weight is less than the capacity: just put the entry to the cache, no "special handling".
* Otherwise do the following while holding an exclusive lock:
  * Explicitly trigger cache cleanup
  * If the cache-weight is less than the admitted capacity (overshooting), put the entry into the cache.
  * Else, reject and update meters accordingly.

We have to allow the "overshooting" to happen to enable (and trigger) the cache cleanup. Otherwise, cleanup would never happen ...