Publish/pom: prevent including test-fixtures dependencies as runtime
The list of dependencies in pom's includes the api/runtime elements of the test-fixtures, which is not what should be published, as it "pulls up" deps like junit, mockito and assertj as Maven runtime scope dependencies.
This change prevents this from happening, although Nessie pom's are fine and do not include test-fixtures dependencies.